move footer back to container so that we can handle the state better
