+++ app/js/containers/Seller/SellerStartContainer.jsx
@@ -1,17 +1,28 @@
 import React, {Component, Fragment} from 'react';
 import PropTypes from 'prop-types';
 import SellerAssets from '../../components/Seller/SellerAssets';
+import Footer from "../../components/Footer";
+
+const assets = ['ETH', 'SNT'];
 
 class SellerStartContainer extends Component {
-  selectAsset = (_selectedAsset) => {
+  constructor(props) {
+    super(props);
+    this.state = {
+      selectedAsset: null
+    };
+  }
+
+  selectAsset = (selectedAsset) => {
+    this.setState({selectedAsset});
     // TODO Save selected asset;
-    this.props.wizard.setReady(true);
   };
 
   render() {
     return (
       <Fragment>
-        <SellerAssets selectAsset={this.selectAsset}/>
+        <SellerAssets selectAsset={this.selectAsset} selectedAsset={this.state.selectedAsset} assets={assets}/>
+        {<Footer next={this.props.wizard.next} ready={this.state.selectedAsset !== null}/>}
       </Fragment>
     );
   }
